+++ repos/base-foc/src/core/platform_thread.cc
@@ -211,7 +211,7 @@ Foc_thread_state Platform_thread::state()
 		s = _pager_obj->state.state;
 
 	s.kcap = _gate.remote;
-	s.id   = _gate.local.local_name();
+	s.id   = (uint16_t)_gate.local.local_name();
 	s.utcb = _utcb;
 
 	return s;
