+++ repos/base-okl4/src/core/platform.cc
@@ -40,13 +40,11 @@ using namespace Genode;
  ** Support for core memory management **
  ****************************************/
 
-bool Mapped_mem_allocator::_map_local(addr_t virt_addr, addr_t phys_addr,
-                                      unsigned size) {
+bool Mapped_mem_allocator::_map_local(addr_t virt_addr, addr_t phys_addr, size_t size) {
 	return map_local(phys_addr, virt_addr, size / get_page_size()); }
 
 
-bool Mapped_mem_allocator::_unmap_local(addr_t virt_addr, addr_t,
-                                        unsigned size) {
+bool Mapped_mem_allocator::_unmap_local(addr_t virt_addr, addr_t, size_t size) {
 	return unmap_local(virt_addr, size / get_page_size()); }
