+++ repos/base/include/cpu_session/cpu_session.h
@@ -210,7 +210,7 @@ struct Genode::Cpu_session : Session
 	 */
 	template<typename T = size_t>
 	static size_t quota_lim_downscale(size_t const value, size_t const limit) {
-		return ((T)value * limit) >> Cpu_session::QUOTA_LIMIT_LOG2; }
+		return (size_t)((T)value * limit) >> Cpu_session::QUOTA_LIMIT_LOG2; }
 
 
 	/*****************************************
